+++ src/test/index.test.ts
@@ -1,4 +1,4 @@
-import anyTest, { TestInterface } from 'ava'
+import anyTest, { TestInterface, ExecutionContext } from 'ava'
 import sinon from 'sinon'
 
 import LevelLogger, { LogTags } from '../index'
@@ -6,6 +6,13 @@ import LevelLogger, { LogTags } from '../index'
 const test = anyTest as TestInterface<{
   consoleInfoStub: sinon.SinonStub<[message?: any, ...optionalParams: any[]], void>
 }>
+
+function calledOnceWithExactly<T, S extends any[]> (
+  t: ExecutionContext<T>, stub: sinon.SinonStub<S, void>, ...args: any[]
+): void {
+  t.true(stub.calledOnce)
+  t.deepEqual(stub.firstCall.args, args)
+}
 
 test.beforeEach(t => {
   t.context.consoleInfoStub = sinon.stub(console, 'info')
@@ -15,16 +22,22 @@ test.afterEach.always(() => {
   sinon.restore()
 })
 
-test.serial('should log correctly with single message', async (t) => {
+test.serial('should log correctly with no parameter', async (t) => {
   const logger = new LevelLogger()
-  logger.info('Hello world!')
-  t.true(t.context.consoleInfoStub.calledOnceWithExactly('Hello world!'))
+  logger.info()
+  calledOnceWithExactly(t, t.context.consoleInfoStub, '')
 })
 
-test.serial('should log correctly with message and parameters', async (t) => {
+test.serial('should log correctly with a single parameter', async (t) => {
+  const logger = new LevelLogger()
+  logger.info('Hello world!')
+  calledOnceWithExactly(t, t.context.consoleInfoStub, 'Hello world!')
+})
+
+test.serial('should log correctly with multiple parameters', async (t) => {
   const logger = new LevelLogger()
   logger.info('Hello', 'again,', 'world!')
-  t.true(t.context.consoleInfoStub.calledOnceWithExactly('Hello again, world!'))
+  calledOnceWithExactly(t, t.context.consoleInfoStub, 'Hello again, world!')
 })
 
 test.serial('should resolve timestamp prefix', async (t) => {
@@ -56,7 +69,7 @@ test.serial('should resolve log level prefix', async (t) => {
     prefixes: [LogTags.MESSAGE_LEVEL]
   })
   logger.info()
-  t.true(t.context.consoleInfoStub.calledOnceWithExactly('INFO'))
+  calledOnceWithExactly(t, t.context.consoleInfoStub, 'INFO')
 })
 
 test.serial('extend should retain previous logger options', async (t) => {
@@ -79,6 +92,6 @@ test.serial('extend should allow change of prefixes', async (t) => {
   })
   logger.info()
   t.true(t.context.consoleInfoStub.calledTwice)
-  t.true(t.context.consoleInfoStub.firstCall.calledWithExactly('A'))
-  t.true(t.context.consoleInfoStub.secondCall.calledWithExactly('B'))
+  t.deepEqual(t.context.consoleInfoStub.firstCall.args, ['A'])
+  t.deepEqual(t.context.consoleInfoStub.secondCall.args, ['B'])
 })