+++ changelog.md
@@ -13,6 +13,7 @@ and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 - Breaking: Split formatting role of option `logger` as separated option `messageFormatter`.
 - Breaking: `LogTags.LOG_LEVEL` renamed as `LogTags.MESSAGE_LEVEL` to distinct logging level from message level.
 - Breaking: Logging level `disabled` renamed to `off` to align with well-known log4j.
+- Breaking: Final logged message was different on edge cases when `printf`-like format string was used as prefix.
 
 ### Removed
 - Breaking: Method `.setLevel()`, setter `.level` and setter `.prefixes` are removed. Use `.extend()` instead.
@@ -21,6 +22,7 @@ and this module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fix prefixes can be altered via `options.prefixes`.
+- Fix incorrect message content was logged in some cases, such as no prefixes and no message parameters supplied.
 
 ## [0.1.2] - 2020-09-29
 ### Fixed